# Remove active_support dependency for the CLI