--- a/archfiend.gemspec
+++ b/archfiend.gemspec
@@ -20,7 +20,6 @@ Gem::Specification.new do |spec|
   spec.executables   = spec.files.grep(%r{^exe/}) { |f| File.basename(f) }
   spec.require_paths = ['lib']
 
-  spec.add_dependency 'activesupport', '~> 5' # CLI
   spec.add_development_dependency 'activerecord', '~> 5' # For specs
   spec.add_development_dependency 'bundler', '~> 1.15'
   spec.add_development_dependency 'config', '~> 1.7'