--- a/lib/archfiend/generators/daemon.rb
+++ b/lib/archfiend/generators/daemon.rb
@@ -1,4 +1,3 @@
-require 'active_support/inflector'
 require 'thor'
 
 module Archfiend