Remove active_support dependency for the CLI
It was only used for `camelize` and `underscore`, this PR adds those methods.
This results in a 400ms (60%) CLI startup speedup, and some CLI RSS reduction.

#### after

```bash
 name                                            time   RSS after
----------------------------------------------------------------
  rubygems/bundler_version_finder                  1ms   5325KB
  thor/command                                     1ms   6070KB
  thor/core_ext/hash_with_indifferent_access       1ms   6103KB
  thor/core_ext/ordered_hash                       1ms   6134KB
  thor/error                                       1ms   6137KB
  thor/invocation                                  1ms   6224KB
  thor/parser/argument                             1ms   6249KB
  thor/parser/arguments                            1ms   6298KB
  thor/parser/option                               1ms   6358KB
  thor/parser/options                              1ms   6413KB
  thor/parser                                     17ms   6413KB
  thor/shell                                       1ms   6431KB
  thor/line_editor/basic                           0ms   6440KB
  readline                                         1ms   6820KB
  thor/line_editor/readline                        2ms   6820KB
  thor/line_editor                                 7ms   6820KB
  thor/util                                        2ms   6890KB
  thor/base                                       77ms   6890KB
  thor                                           147ms   6890KB
  pathname.so                                      1ms   7045KB
  pathname                                         3ms   7045KB
  forwardable/impl                                 0ms   7081KB
  forwardable                                      2ms   7081KB
  archfiend/version                                0ms   7087KB
  thor/group                                       2ms   7162KB
  thor/core_ext/io_binary_read                     0ms   7218KB
  thor/actions/empty_directory                     1ms   7256KB
  thor/actions/create_file                         2ms   7256KB
  thor/actions/create_link                         1ms   7264KB
  thor/actions/directory                           1ms   7282KB
  cgi/escape                                       1ms   7569KB
  cgi/util                                         2ms   7569KB
  strscan                                          1ms   7594KB
  erb                                             14ms   7594KB
  thor/actions/file_manipulation                  15ms   7594KB
  thor/actions/inject_into_file                    1ms   7606KB
  thor/actions                                    54ms   7606KB
  archfiend/generators/daemon                     61ms   7606KB
  archfiend/generators/options                     1ms   7658KB
  archfiend/generators/extensions                  1ms   7686KB
  archfiend/generators/utils                       1ms   7693KB
  archfiend/core_ext/string/camelize               0ms   7694KB
  archfiend/core_ext/string/underscore             0ms   7700KB
----------------------------------------------------------------
./lib/archfiend/cli                              276ms   7700KB
```
#### before

```bash
  name                                                      time   RSS after
--------------------------------------------------------------------------
  rubygems/bundler_version_finder                            1ms   5346KB
  thor/command                                               1ms   6092KB
  thor/core_ext/hash_with_indifferent_access                 1ms   6118KB
  thor/core_ext/ordered_hash                                 1ms   6186KB
  thor/error                                                 0ms   6191KB
  thor/invocation                                            1ms   6257KB
  thor/parser/argument                                       1ms   6277KB
  thor/parser/arguments                                      1ms   6318KB
  thor/parser/option                                         1ms   6381KB
  thor/parser/options                                        1ms   6460KB
  thor/parser                                               17ms   6460KB
  thor/shell                                                 1ms   6488KB
  thor/line_editor/basic                                     0ms   6504KB
  readline                                                   1ms   6889KB
  thor/line_editor/readline                                  2ms   6889KB
  thor/line_editor                                           7ms   6889KB
  thor/util                                                  1ms   6968KB
  thor/base                                                 76ms   6968KB
  thor                                                     143ms   6968KB
  pathname.so                                                1ms   7127KB
  pathname                                                   3ms   7127KB
  forwardable/impl                                           0ms   7160KB
  forwardable                                                1ms   7160KB
  archfiend/version                                          0ms   7165KB
  concurrent/constants                                       0ms   7536KB
  concurrent/utility/engine                                  1ms   7548KB
  concurrent/synchronization/abstract_object                 1ms   7549KB
  concurrent/utility/native_extension_loader                 1ms   7556KB
  concurrent/synchronization/mri_object                      1ms   7836KB
  concurrent/synchronization/jruby_object                    1ms   7841KB
  concurrent/synchronization/rbx_object                      1ms   7843KB
  concurrent/synchronization/truffle_object                  1ms   7852KB
  concurrent/synchronization/object                          1ms   7867KB
  concurrent/synchronization/volatile                        1ms   7883KB
  concurrent/synchronization/abstract_lockable_object        1ms   7884KB
  concurrent/synchronization/mri_lockable_object             1ms   7895KB
  concurrent/synchronization/jruby_lockable_object           1ms   7896KB
  concurrent/synchronization/rbx_lockable_object             1ms   7900KB
  concurrent/synchronization/truffle_lockable_object         1ms   7911KB
  concurrent/synchronization/lockable_object                 1ms   7919KB
  concurrent/synchronization/condition                       1ms   7926KB
  concurrent/synchronization/lock                            1ms   7928KB
  concurrent/synchronization                               111ms   7928KB
  concurrent/collection/map/non_concurrent_map_backend       1ms   7945KB
  concurrent/collection/map/mri_map_backend                  3ms   7945KB
  concurrent/map                                           130ms   7945KB
  active_support/core_ext/array/prepend_and_append           1ms   7945KB
  active_support/core_ext/regexp                             1ms   7946KB
  active_support/core_ext/hash/deep_merge                    1ms   7950KB
  active_support/core_ext/hash/except                        1ms   7952KB
  active_support/core_ext/hash/slice                         1ms   7954KB
  i18n/version                                               1ms   8004KB
  cgi/core                                                   4ms   8298KB
  cgi/escape                                                 1ms   8368KB
  cgi/util                                                   4ms   8368KB
  cgi/cookie                                                 6ms   8368KB
  cgi                                                       28ms   8380KB
  i18n/exceptions                                           29ms   8380KB
  i18n/interpolate/ruby                                      1ms   8390KB
  i18n                                                      46ms   8390KB
  active_support/lazy_load_hooks                             1ms   8394KB
  i18n/config                                                2ms   8405KB
  active_support/i18n                                       75ms   8405KB
  singleton                                                  2ms   8410KB
  active_support/core_ext/kernel/singleton_class             1ms   8413KB
  active_support/core_ext/module/delegation                  3ms   8436KB
  active_support/deprecation/instance_delegator              8ms   8436KB
  securerandom                                               2ms   8468KB
  active_support/notifications/instrumenter                  3ms   8468KB
  mutex_m                                                    2ms   8491KB
  active_support/notifications/fanout                        7ms   8498KB
  active_support/per_thread_registry                         1ms   8510KB
  active_support/notifications                              24ms   8510KB
  active_support/deprecation/behaviors                      25ms   8510KB
  active_support/deprecation/reporting                       3ms   8546KB
  active_support/deprecation/constant_accessor               1ms   8558KB
  active_support/core_ext/module/aliasing                    1ms   8596KB
  active_support/core_ext/array/extract_options              1ms   8609KB
  active_support/deprecation/method_wrappers                 6ms   8609KB
  active_support/deprecation/proxy_wrappers                  1ms   8651KB
  active_support/core_ext/module/deprecation                 1ms   8746KB
  active_support/deprecation                                91ms   8746KB
  active_support/inflector/inflections                     326ms   8746KB
  active_support/multibyte                                   1ms   8790KB
  active_support/core_ext/string/multibyte                   1ms   8790KB
  active_support/inflector/transliterate                    10ms   8800KB
  active_support/inflections                                 2ms   8962KB
  active_support/inflector/methods                           8ms   8964KB
  active_support/core_ext/string/inflections                 1ms   8985KB
  active_support/inflector                                 378ms   8985KB
  thor/group                                                 2ms   9045KB
  thor/core_ext/io_binary_read                               1ms   9099KB
  thor/actions/empty_directory                               1ms   9130KB
  thor/actions/create_file                                   3ms   9130KB
  thor/actions/create_link                                   1ms   9147KB
  thor/actions/directory                                     2ms   9174KB
  strscan                                                    2ms   9532KB
  erb                                                        7ms   9532KB
  thor/actions/file_manipulation                             9ms   9532KB
  thor/actions/inject_into_file                              2ms   9563KB
  thor/actions                                              51ms   9563KB
  archfiend/generators/daemon                              445ms   9563KB
  archfiend/generators/options                               4ms   9635KB
  archfiend/generators/extensions                            2ms   9680KB
  archfiend/generators/utils                                 2ms   9692KB
  archfiend/core_ext/string/camelize                         1ms   9711KB
  archfiend/core_ext/string/underscore                       1ms   9734KB
--------------------------------------------------------------------------
./lib/archfiend/cli                                        662ms   9734KB
```


Before submitting the PR make sure the following are checked:

* [ ] Wrote [good commit messages][1].
* [ ] Commit message starts with `[Fix #issue-number]` (if the related issue exists).
* [ ] Feature branch is up-to-date with `master` (if not - rebase it).
* [ ] Squashed related commits together.
* [ ] Added tests.
* [ ] Added an entry to the [Changelog](../blob/master/CHANGELOG.md) if the new code introduces user-observable changes. See [changelog entry format](../blob/master/CONTRIBUTING.md#changelog-entry-format).
* [ ] The PR relates to *only* one subject with a clear title
  and description in grammatically correct, complete sentences.
* [ ] Run `bundle exec rspec` & `bundle exec rubocop`.

[1]: http://chris.beams.io/posts/git-commit/